@@ -3,7 +3,7 @@ let worker: Worker | undefined = undefined;
 export const useWorker = () => {
     try {
         if (!worker)
-            worker = new Worker('../worker/pixelMapper.worker.ts', { type: 'module' })
+            worker = new Worker(new URL('../worker/pixelMapper.worker.ts', import.meta.url));
         return worker;
     } catch {
         return undefined
